requirements-lib.txt
eciespy
httpx
jinja2
joblib==1.1.0
PyNaCl==1.5.0
python-multipart
numpy==1.21.0
starlette
scikit-learn==1.0.1
uvicorn
web3
websockets
